+++ index.js
@@ -47,6 +47,9 @@ const ITREXFactory = require('./artifacts/contracts/factory/ITREXFactory.sol/ITR
 const TREXFactory = require('./artifacts/contracts/factory/TREXFactory.sol/TREXFactory.json');
 // DVD
 const DVDTransferManager = require('./artifacts/contracts/DVD/DVDTransferManager.sol/DVDTransferManager.json');
+// DVA
+const DVATransferManager = require('./artifacts/contracts/DVA/DVATransferManager.sol/DVATransferManager.json');
+const IDVATransferManager = require('./artifacts/contracts/DVA/IDVATransferManager.sol/IDVATransferManager.json');
 // compliance
 const IModularCompliance = require('./artifacts/contracts/compliance/modular/IModularCompliance.sol/IModularCompliance.json');
 const MCStorage = require('./artifacts/contracts/compliance/modular/MCStorage.sol/MCStorage.json');
@@ -106,6 +109,8 @@ module.exports = {
     TREXFactory,
     // DVD
     DVDTransferManager,
+    // DVA
+    DVATransferManager,
     // compliance
     MCStorage,
     ModularCompliance,
@@ -135,5 +140,6 @@ module.exports = {
     ITREXFactory,
     IModularCompliance,
     IModule,
+    IDVATransferManager,
   },
 };